Conflicts:
	scilab/Makefile.in
	scilab/Scilab.sln
	scilab/bin/scilab
	scilab/configure
	scilab/configure.ac
	scilab/contrib/Makefile.in
	scilab/etc/modules.xml.in
	scilab/libs/LibScilab/LibScilab.vcproj
	scilab/modules/Makefile.am
	scilab/modules/api_scilab/src/cpp/api_common.cpp
	scilab/modules/boolean/Makefile.in
	scilab/modules/call_scilab/src/c/call_scilab.c
	scilab/modules/commons/Makefile.am
	scilab/modules/commons/Makefile.in
	scilab/modules/core/Makefile.am
	scilab/modules/core/Makefile.in
	scilab/modules/core/includes/gw_core.h
	scilab/modules/core/sci_gateway/c/sci_exit.c
	scilab/modules/core/src/c/SCIHOME.c
	scilab/modules/core/src/c/banier.c
	scilab/modules/core/src/c/core.vcproj
	scilab/modules/core/src/c/getmodules.c
	scilab/modules/double/Makefile.in
	scilab/modules/elementary_functions/Makefile.in
	scilab/modules/elementary_functions/macros/dec2oct.sci
	scilab/modules/elementary_functions/sci_gateway/c/gw_elementary_functions.c
	scilab/modules/elementary_functions/sci_gateway/c/sci_maxi.c
	scilab/modules/elementary_functions/sci_gateway/elementary_functions_gateway.xml
	scilab/modules/fileio/Makefile.in
	scilab/modules/fileio/macros/mkdir.sci
	scilab/modules/fileio/sci_gateway/c/sci_copyfile.c
	scilab/modules/fileio/sci_gateway/c/sci_fileparts.c
	scilab/modules/fileio/sci_gateway/c/sci_isdir.c
	scilab/modules/fileio/sci_gateway/c/sci_mclose.c
	scilab/modules/fileio/sci_gateway/c/sci_mget.c
	scilab/modules/fileio/sci_gateway/c/sci_mput.c
	scilab/modules/fileio/sci_gateway/c/sci_mputl.c
	scilab/modules/fileio/sci_gateway/c/sci_pathconvert.c
	scilab/modules/fileio/src/c/copyfile_others.c
	scilab/modules/fileio/src/c/fprintfMat.c
	scilab/modules/fileio/src/c/fscanfMat.c
	scilab/modules/fileio/src/c/mgetl.c
	scilab/modules/functions/Makefile.in
	scilab/modules/graph/src/java/org/scilab/modules/graph/io/ScilabIntegerCodec.java
	scilab/modules/gui/Makefile.am
	scilab/modules/gui/Makefile.in
	scilab/modules/gui/sci_gateway/c/sci_ClipBoard.c
	scilab/modules/gui/src/java/org/scilab/modules/gui/bridge/window/SwingScilabWindow.java
	scilab/modules/hdf5/sci_gateway/cpp/sci_export_to_hdf5.cpp
	scilab/modules/hdf5/sci_gateway/cpp/sci_import_from_hdf5.cpp
	scilab/modules/integer/Makefile.in
	scilab/modules/io/Makefile.in
	scilab/modules/io/sci_gateway/c/sci_file.c
	scilab/modules/javasci/src/c/javasci_Scilab.c
	scilab/modules/localization/src/c/setgetlanguage.c
	scilab/modules/matio/src/c/CreateHyperMatrixVariable.c
	scilab/modules/matio/src/c/GetMlistVariable.c
	scilab/modules/matio/src/c/GetStructVariable.c
	scilab/modules/output_stream/output_stream.iss
	scilab/modules/output_stream/sci_gateway/c/sci_mprintf.c
	scilab/modules/output_stream/sci_gateway/c/sci_msprintf.c
	scilab/modules/output_stream/src/c/api_scilab_Import.def
	scilab/modules/output_stream/src/c/errmsg.c
	scilab/modules/pvm/Makefile.in
	scilab/modules/scicos/Makefile.in
	scilab/modules/scicos_blocks/Makefile.am
	scilab/modules/scicos_blocks/Makefile.in
	scilab/modules/scinotes/Makefile.am
	scilab/modules/scinotes/Makefile.in
	scilab/modules/shell/Makefile.in
	scilab/modules/special_functions/sci_gateway/c/sci_oldbesseli.c
	scilab/modules/special_functions/sci_gateway/c/sci_oldbesselj.c
	scilab/modules/special_functions/sci_gateway/c/sci_oldbesselk.c
	scilab/modules/special_functions/sci_gateway/c/sci_oldbessely.c
	scilab/modules/string/Makefile.in
	scilab/modules/string/sci_gateway/c/sci_grep.c
	scilab/modules/string/src/c/pcre_private.h
	scilab/modules/string/src/c/stringsstrrchr.c
	scilab/modules/time/Makefile.in
	scilab/modules/time/sci_gateway/c/sci_calendar.c
	scilab/modules/time/time.iss
	scilab/modules/types/Makefile.am
	scilab/modules/types/Makefile.in
	scilab/modules/types/readme.txt
	scilab/modules/ui_data/Makefile.am
	scilab/modules/ui_data/Makefile.in
	scilab/modules/windows_tools/sci_gateway/c/sci_win64.c

+++ scilab/modules/fileio/src/c/fprintfMat.c
@@ -1,11 +1,11 @@
 /*
 * Scilab ( http://www.scilab.org/ ) - This file is part of Scilab
 * Copyright (C) 2010 - DIGITEO - Allan CORNET
-* 
+*
 * This file must be used under the terms of the CeCILL.
 * This source file is licensed as described in the file COPYING, which
 * you should have received as part of this distribution.  The terms
-* are also available at    
+* are also available at
 * http://www.cecill.info/licences/Licence_CeCILL_V2-en.txt
 *
 */
@@ -27,15 +27,15 @@
 
 /*--------------------------------------------------------------------------*/
 #ifdef _MSC_VER
-#define finite(x) _finite(x) 
+#define finite(x) _finite(x)
 #endif
 #define EOL "\n"
 #define NanString "Nan"
 #define InfString "Inf"
 #define NegInfString "-Inf"
 /*--------------------------------------------------------------------------*/
-#define NB_FORMAT_SUPPORTED 7 
-static char *supportedFormat[NB_FORMAT_SUPPORTED] = 
+#define NB_FORMAT_SUPPORTED 7
+static char *supportedFormat[NB_FORMAT_SUPPORTED] =
 {"lf", "lg", "d", "i", "e", "f", "g"};
 /*--------------------------------------------------------------------------*/
 static BOOL checkFprintfMatFormat(char *format);
@@ -57,7 +57,7 @@ static int signbit(double x)
     return u.i[1] < 0;
 #else
     return u.s[3] < 0;
-#endif 
+#endif
 }
 #endif /* signbit */
 /*--------------------------------------------------------------------------*/
@@ -89,7 +89,7 @@ fprintfMatError fprintfMat(char *filename, char *format, char *separator,
     {
         if (textAdded)
         {
-            for (i = 0; i < sizeTextAdded; i++) 
+            for (i = 0; i < sizeTextAdded; i++)
             {
                 if (textAdded[i])
                 {
@@ -218,7 +218,7 @@ static char *getCleanedFormat(char *format)
                 char *token = strstr(percent, supportedFormat[i]);
                 if (token)
                 {
-                    int nbcharacters = strlen(percent) - strlen(token);          
+                    int nbcharacters = (int)(strlen(percent) - strlen(token));
                     cleanedFormat = os_strdup(percent);
                     cleanedFormat[nbcharacters] = 0;
                     if ( ((nbcharacters - 1 > 0) && (isdigit(cleanedFormat[nbcharacters-1])) ||
